% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gleamTest.R
\name{gleamTest}
\alias{gleamTest}
\title{Perform GLEAM test}
\usage{
gleamTest(
  query,
  subject,
  txdb,
  background = c("genome", "gene", "exon", "intron"),
  ...
)
}
\arguments{
\item{query}{A vector of bedpe files or a list of genomic interaction data
(\link[S4Vectors:Pairs-class]{Pairs} or
\link[InteractionSet:GInteractions-class]{GInteractions})}

\item{subject}{A vector of bed files or a GRangesList object. Optional.}

\item{txdb}{An TxDb object to retrieve the sequence lengths, and genes
annotations. Or an GRanges object.}

\item{background}{Background mode. The test will restricted within the
region.}

\item{\dots}{parameters used by
\link[InteractionSet:findOverlaps]{findOverlaps}.}
}
\description{
Run Genomic Loops Enrichment Analysis Method test
}
\examples{
# example code
pd <- system.file("extdata", package = "hicVennDiagram", mustWork = TRUE)
fs <- dir(pd, pattern = ".bedpe", full.names = TRUE)
library(TxDb.Hsapiens.UCSC.hg38.knownGene)
gleamTest(fs, txdb=TxDb.Hsapiens.UCSC.hg38.knownGene)
gr <- exons(TxDb.Hsapiens.UCSC.hg38.knownGene)
grl <- GRangesList(gr1=gr, gr2=gr, gr3=gr)
gleamTest(fs, grl, txdb=TxDb.Hsapiens.UCSC.hg38.knownGene,
 background = 'gene')
}
