% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gleamTest.R
\name{gleamTest}
\alias{gleamTest}
\title{Perform GLEAM test}
\usage{
gleamTest(
  query,
  subject,
  txdb,
  background = c("genome", "gene", "exon", "intron"),
  ...
)
}
\arguments{
\item{query, subject}{A vector of bedpe files or a list of genomic
interaction data (\link[S4Vectors:Pairs-class]{Pairs} or
\link[InteractionSet:GInteractions-class]{GInteractions}) or a list of
GRanges object. 'subject' is optional if length of query > 1.}

\item{txdb}{An TxDb object to retrieve the sequence lengths, and genes
annotations. Or an GRanges object.}

\item{background}{Background mode. The test will restricted within the
region.}

\item{\dots}{parameters used by
\link[InteractionSet:findOverlaps]{findOverlaps}.}
}
\description{
Run Genomic Loops Enrichment Analysis Method test
}
\examples{
# example code
pd <- system.file("extdata", package = "hicVennDiagram", mustWork = TRUE)
fs <- dir(pd, pattern = ".bedpe", full.names = TRUE)
library(TxDb.Hsapiens.UCSC.hg38.knownGene)
gleamTest(fs, txdb=TxDb.Hsapiens.UCSC.hg38.knownGene)
grl <- GRangesList(exons=reduce(exons(TxDb.Hsapiens.UCSC.hg38.knownGene)),
 genes=reduce(genes(TxDb.Hsapiens.UCSC.hg38.knownGene)))
gleamTest(fs[seq_along(grl)], grl, txdb=TxDb.Hsapiens.UCSC.hg38.knownGene,
 background = 'gene')
gleamTest(grl, fs[seq_along(grl)], txdb=TxDb.Hsapiens.UCSC.hg38.knownGene,
 background = 'gene')
gleamTest(grl[c(2, 1)], grl, txdb=TxDb.Hsapiens.UCSC.hg38.knownGene,
 background = 'gene')
gleamTest(grl, txdb=TxDb.Hsapiens.UCSC.hg38.knownGene,
 background = 'gene')
gleamTest(grl, txdb=TxDb.Hsapiens.UCSC.hg38.knownGene,
 background = 'genome')
}
