% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsetPlot.R
\name{upsetPlot}
\alias{upsetPlot}
\title{UpSet plot for the Venn count table}
\usage{
upsetPlot(
  vennTable,
  label_all = list(na.rm = TRUE, color = "gray30", alpha = 0.7, label.padding = unit(0.1,
    "lines")),
  coln_prefix = NULL
)
}
\arguments{
\item{vennTable}{An vennTable object, the first element in the output of
\link{vennCount}.}

\item{label_all}{A list of parameters used by \link[ggplot2]{geom_label} for
text labels of counts for each group. If it set to FALSE or length of the
list is zero, the labels will be ignored.}

\item{coln_prefix}{The prefix to be removed for colnumn names of vennTable.}
}
\value{
A ggplot object.
}
\description{
Plot the overlaps counts by ComplexUpset.
}
\examples{
pd <- system.file("extdata", package = "hicVennDiagram", mustWork = TRUE)
fs <- dir(pd, pattern = ".bedpe", full.names = TRUE)
vc <- vennCount(fs)
upsetPlot(vc)

}
