% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vennPlot.R
\name{vennPlot}
\alias{vennPlot}
\title{Venn diagram for the Venn count table}
\usage{
vennPlot(vennTable)
}
\arguments{
\item{vennTable}{An vennTable object, the first element in the output of
\link{vennCount}.}
}
\value{
A grid object.
}
\description{
Plot the overlaps counts by eulerr.
}
\examples{
pd <- system.file("extdata", package = "hicVennDiagram", mustWork = TRUE)
fs <- dir(pd, pattern = ".bedpe", full.names = TRUE)
vc <- vennCount(fs)
vennPlot(vc)
}
